String value only allow Blank + migrations
